requirements.txt
altair==4.1.0
aniso8601==9.0.1
argon2-cffi==20.1.0
astor==0.8.1
attrs==21.2.0
autopep8==1.5.7
backcall==0.2.0
base58==2.1.0
bleach==4.1.0
blinker==1.4
blis==0.7.4
cachetools==4.2.2
catalogue==2.0.6
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.4
cycler==0.10.0
cymem==2.0.5
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
Flask==2.0.1
Flask-RESTful==0.3.9
gitdb==4.0.7
GitPython==3.1.18
idna==3.2
install==1.3.4
ipykernel==6.1.0
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
murmurhash==1.0.5
mysql-connector==2.2.9
mysql-connector-python==8.0.26
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.3
numpy==1.21.1
packaging==21.0
pandas==1.3.1
pandocfilters==1.4.3
parso==0.8.2
pathy==0.6.0
pickleshare==0.7.5
Pillow==8.3.1
preshed==3.0.5
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
pyarrow==5.0.0
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
pydeck==0.6.2
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
pywin32==301
pywinpty==1.1.3
pyzmq==22.2.1
requests==2.26.0
scipy==1.7.1
seaborn==0.11.1
Send2Trash==1.8.0
six==1.16.0
smart-open==5.2.0
smmap==4.0.0
spacy==3.1.2
spacy-legacy==3.0.8
spacy-streamlit==1.0.2
srsly==2.4.1
streamlit==0.87.0
terminado==0.11.1
testpath==0.5.0
thinc==8.0.8
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.62.2
traitlets==5.0.5
typer==0.3.2
typing-extensions==3.10.0.0
tzdata==2021.1
tzlocal==3.0
urllib3==1.26.6
validators==0.18.2
wasabi==0.8.2
watchdog==2.1.5
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
